fix: renamed package and update logic init
